@@ -1,11 +1,13 @@
 package fr.robot.robottags.manager;
 
-import fr.robot.robottags.Main;
 import fr.robot.robottags.object.Tag;
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.HashMap;
 import java.util.UUID;
+
+import static fr.robot.robottags.Main.debug;
 
 public class PlayerManager {
 
@@ -24,54 +26,29 @@ public class PlayerManager {
     }
 
     public static void load(UUID playerUUID) {
-        Main.sendDebug("Chargement du tag pour un joueur. (" + playerUUID.toString() + ")");
-
-        String tagID = null;
-        switch (StorageManager.getMode()) {
-            case YML:
-                tagID = ConfigManager.getDatabase().get().getString(playerUUID.toString());
-                break;
-            case MYSQL:
-                tagID = MysqlManager.getTag(playerUUID);
-                break;
-        }
+        debug("Chargement du tag pour un joueur. (" + playerUUID.toString() + ")");
+        String tagID = StorageManager.getData().getTagId(playerUUID);
 
         if(tagID != null && TagManager.exist(tagID)) {
             setTag(playerUUID, tagID);
-            Main.sendDebug("Le joueur possède le tag " + tagID + " (" + StorageManager.getMode() + ")");
+            debug("Le joueur possède le tag " + tagID + " (" + StorageManager.getMode() + ")");
         } else {
-            Main.sendDebug("Le joueur n'a été vu avec aucun tag existant." +  " (" + StorageManager.getMode() + ")");
+            debug("Le joueur n'a été vu avec aucun tag existant." +  " (" + StorageManager.getMode() + ")");
         }
     }
 
     public static void save(UUID playerUUID) {
         String tagID = getTagId(playerUUID);
         if(tagID == null) {
-            Main.sendDebug("Le joueur avec l'UUID " + playerUUID + " ne possède pas de tag.");
+            debug("Le joueur avec l'UUID " + playerUUID + " ne possède pas de tag.");
             return;
         }
-        switch (StorageManager.getMode()) {
-            case YML:
-                ConfigManager.getDatabase().get().set(playerUUID.toString(), tagID);
-                ConfigManager.getDatabase().save();
-                break;
-            case MYSQL:
-                MysqlManager.setTag(playerUUID, tagID);
-                break;
-        }
-        Main.sendDebug("Le joueur avec l'UUID " + playerUUID + " a été sauvegardé avec le tag " + tagID + " + (" + StorageManager.getMode() + ")");
+        StorageManager.getData().setTagId(playerUUID, tagID);
+        debug("Le joueur avec l'UUID " + playerUUID + " a été sauvegardé avec le tag " + tagID + " + (" + StorageManager.getMode() + ")");
     }
 
     public static void clear(UUID playerUUID) {
-        switch (StorageManager.getMode()) {
-            case YML:
-                ConfigManager.getDatabase().get().set(playerUUID.toString(), null);
-                ConfigManager.getDatabase().save();
-                break;
-            case MYSQL:
-                MysqlManager.setTag(playerUUID, null);
-                break;
-        }
+        StorageManager.getData().setTagId(playerUUID, null);
         playerTags.remove(playerUUID);
     }
 
@@ -79,7 +56,7 @@ public class PlayerManager {
         save(player.getUniqueId());
     }
 
-    public static String getTagId(UUID playerUUID) {
+    public static @Nullable String getTagId(UUID playerUUID) {
         if(playerTags.containsKey(playerUUID) && TagManager.exist(playerTags.get(playerUUID)))
             return playerTags.get(playerUUID);
         return null;
